Use doubles for time in many other places

+++ editor/plugins/canvas_item_editor_plugin.h
@@ -535,7 +535,7 @@ private:
 	HSplitContainer *palette_split;
 	VSplitContainer *bottom_split;
 
-	void _popup_warning_temporarily(Control *p_control, const float p_duration);
+	void _popup_warning_temporarily(Control *p_control, const double p_duration);
 	void _popup_warning_depop(Control *p_control);
 
 	void _set_owner_for_node_and_children(Node *p_node, Node *p_owner);
